+++ Source/Bugsnag.m
@@ -33,6 +33,11 @@
 
 static BugsnagClient *bsg_g_bugsnag_client = NULL;
 
+@interface BugsnagConfiguration ()
+@property(readwrite, retain, nullable) BugsnagMetadata *metadata;
+@property(readwrite, retain, nullable) BugsnagMetadata *config;
+@end
+
 @interface Bugsnag ()
 + (BugsnagClient *)client;
 + (BOOL)bugsnagStarted;
@@ -49,8 +54,7 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
 @implementation Bugsnag
 
 + (void)startBugsnagWithApiKey:(NSString *)apiKey {
-    BugsnagConfiguration *configuration = [BugsnagConfiguration new];
-    configuration.apiKey = apiKey;
+    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:apiKey];
     [self startBugsnagWithConfiguration:configuration];
 }
 
@@ -125,36 +129,6 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
                                        block(report);
                                    }
                                }];
-    }
-}
-
-+ (void)notify:(NSException *)exception withData:(NSDictionary *)metadata {
-    if ([self bugsnagStarted]) {
-        [[self client]
-                notifyException:exception
-                          block:^(BugsnagEvent *_Nonnull report) {
-                              report.depth += 2;
-                              report.metadata = [metadata
-                                      BSG_mergedInto:[self.client.configuration
-                                              .metadata toDictionary]];
-                          }];
-    }
-}
-
-+ (void)notify:(NSException *)exception
-      withData:(NSDictionary *)metadata
-    atSeverity:(NSString *)severity {
-    if ([self bugsnagStarted]) {
-        [[self client]
-                notifyException:exception
-                     atSeverity:BSGParseSeverity(severity)
-                          block:^(BugsnagEvent *_Nonnull report) {
-                              report.depth += 2;
-                              report.metadata = [metadata
-                                      BSG_mergedInto:[self.client.configuration
-                                              .metadata toDictionary]];
-                              report.severity = BSGParseSeverity(severity);
-                          }];
     }
 }
 
@@ -271,6 +245,7 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
     dispatch_once(&onceToken, ^{
       formatter = [NSDateFormatter new];
       formatter.dateFormat = @"yyyy'-'MM'-'dd'T'HH':'mm':'ssZZZ";
+      formatter.timeZone = [NSTimeZone timeZoneForSecondsFromGMT:0];
     });
     return formatter;
 }
@@ -312,6 +287,13 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
 + (void)removeOnSessionBlock:(BugsnagOnSessionBlock _Nonnull )block
 {
     [[self configuration] removeOnSessionBlock:block];
+}
+
+/**
+ * Intended for internal use only - sets the code bundle id for React Native
+ */
++ (void)updateCodeBundleId:(NSString *)codeBundleId {
+    [self configuration].codeBundleId = codeBundleId;
 }
 
 // =============================================================================