+++ Source/BugsnagClient.m
@@ -69,8 +69,8 @@ struct bugsnag_data_t {
     // Contains properties in the Bugsnag payload overridden by the user before
     // it was sent
     char *userOverridesJSON;
-    // User onError handler
-    void (*onError)(const BSG_KSCrashReportWriter *writer);
+    // User onCrash handler
+    void (*onCrash)(const BSG_KSCrashReportWriter *writer);
 };
 
 static struct bugsnag_data_t bsg_g_bugsnag_data;
@@ -118,7 +118,7 @@ void BSSerializeDataCrashHandler(const BSG_KSCrashReportWriter *writer, int type
         }
         if (crashSentinelPath != NULL) {
             // Create a file to indicate that the crash has been handled by
-            // the library. This exists in case the subsequent `onError` handler
+            // the library. This exists in case the subsequent `onCrash` handler
             // crashes or otherwise corrupts the crash report file.
             int fd = open(crashSentinelPath, O_RDWR | O_CREAT, 0644);
             if (fd > -1) {
@@ -127,8 +127,8 @@ void BSSerializeDataCrashHandler(const BSG_KSCrashReportWriter *writer, int type
         }
     }
 
-    if (bsg_g_bugsnag_data.onError) {
-        bsg_g_bugsnag_data.onError(writer);
+    if (bsg_g_bugsnag_data.onCrash) {
+        bsg_g_bugsnag_data.onCrash(writer);
     }
 }
 
@@ -261,6 +261,10 @@ void BSGWriteSessionCrashData(BugsnagSession *session) {
 
 @interface BugsnagConfiguration ()
 @property(nonatomic, readwrite, strong) NSMutableSet *plugins;
+@property(readonly, retain, nullable) NSURL *notifyURL;
+@property(readwrite, retain, nullable) BugsnagMetadata *metadata;
+@property(readwrite, retain, nullable) BugsnagMetadata *config;
+@property(readonly, strong, nullable) BugsnagBreadcrumbs *breadcrumbs;
 @end
 
 @implementation BugsnagClient
@@ -314,7 +318,7 @@ NSString *_lastOrientation = nil;
         self.crashSentry = [BugsnagCrashSentry new];
         self.errorReportApiClient = [[BugsnagErrorReportApiClient alloc] initWithConfig:configuration
                                                                               queueName:@"Error API queue"];
-        bsg_g_bugsnag_data.onError = (void (*)(const BSG_KSCrashReportWriter *))self.configuration.onError;
+        bsg_g_bugsnag_data.onCrash = (void (*)(const BSG_KSCrashReportWriter *))self.configuration.onCrashHandler;
 
         static dispatch_once_t once_t;
         dispatch_once(&once_t, ^{
@@ -644,17 +648,6 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
                      block(report);
                  }
                }];
-}
-
-- (void)notifyException:(NSException *)exception
-             atSeverity:(BSGSeverity)severity
-                  block:(void (^)(BugsnagEvent *))block {
-
-    BugsnagHandledState *state = [BugsnagHandledState
-        handledStateWithSeverityReason:UserSpecifiedSeverity
-                              severity:severity
-                             attrValue:nil];
-    [self notify:exception handledState:state block:block];
 }
 
 - (void)notifyException:(NSException *)exception