+++ Tests/BugsnagTests.m
@@ -20,6 +20,7 @@
 
 @interface BugsnagConfiguration ()
 @property(nonatomic, readwrite, strong) NSMutableArray *onSendBlocks;
+@property(readwrite, retain, nullable) BugsnagMetadata *metadata;
 @end
 
 @interface BugsnagClient ()